new setSize block for applab
